+++ lib/amazonka/src/Amazonka/Auth/SSO.hs
@@ -17,7 +17,7 @@ import Amazonka.Env (Env, Env' (..))
 import Amazonka.Prelude
 import Amazonka.SSO.GetRoleCredentials as SSO
 import qualified Amazonka.SSO.Types as SSO (RoleCredentials (..))
-import Amazonka.Send (sendUnsigned)
+import Amazonka.Send (sendUnsignedEither)
 import Amazonka.Types
 import Control.Exception (IOException)
 import qualified Control.Exception as Exception
@@ -92,19 +92,26 @@ fromSSO cachedTokenFile ssoRegion accountId roleName env = do
               accountId
               (fromSensitive accessToken)
 
-      resp <- runResourceT $ sendUnsigned ssoEnv getRoleCredentials
-      pure . roleCredentialsToAuthEnv $
-        resp ^. SSO.getRoleCredentialsResponse_roleCredentials
+      runResourceT (sendUnsignedEither ssoEnv getRoleCredentials) >>= \case
+        Left err -> Exception.throwIO (errorAsAuthError err)
+        Right resp ->
+          pure . roleCredentialsToAuthEnv $
+            resp ^. SSO.getRoleCredentialsResponse_roleCredentials
+
+    errorAsAuthError = \case
+      ServiceError err -> AuthServiceError err
+      TransportError err -> RetrievalError err
+      other -> OtherAuthError (Exception.toException other)
 
 -- | Return the cached token file for a given @sso_start_url@
 --
 -- Matches
 -- [botocore](https://github.com/boto/botocore/blob/c02f3561f56085b8a3f98501d25b9857b916c10e/botocore/utils.py#L2596-L2597),
 -- so that we find tokens produced by @aws sso login@.
-relativeCachedTokenFile :: (MonadIO m) => Text -> m FilePath
-relativeCachedTokenFile startUrl = do
-  let sha1 = show . Crypto.hashSHA1 $ Text.encodeUtf8 startUrl
-  pure $ "/.aws/sso/cache/" <> sha1 <> ".json"
+relativeCachedTokenFile :: Text -> FilePath
+relativeCachedTokenFile startUrl = "/.aws/sso/cache/" <> sha1 <> ".json"
+  where
+    sha1 = show . Crypto.hashSHA1 $ Text.encodeUtf8 startUrl
 
 readCachedAccessToken :: (MonadIO m) => FilePath -> m CachedAccessToken
 readCachedAccessToken p = liftIO $